refactor: #29 use client 정리
